Extract out code that creates CacheFeatureValue shared between edge-rest and MR
Honours the percentage override for a shared strategy
