+++ backend/edge-rest/src/main/kotlin/io/featurehub/edge/db/sql/DbDachaCacheSqlApi.kt
@@ -5,15 +5,17 @@ import io.featurehub.dacha.api.DachaClientServiceRegistry
 import io.featurehub.dacha.api.DachaEnvironmentService
 import io.featurehub.dacha.model.DachaStructureResponse
 import io.featurehub.db.publish.CacheSourceFeatureGroupApi
+import io.featurehub.db.publish.CommonCacheGenerator
 import io.featurehub.db.publish.FeatureModelWalker
 import io.featurehub.db.services.Conversions
 import jakarta.inject.Inject
 import java.util.*
 
 class DbDachaCacheSqlApi @Inject constructor(cacheSourceFeatureGroup: CacheSourceFeatureGroupApi,
+                                             commonCacheGenerator: CommonCacheGenerator,
                                              featureModelWalker: FeatureModelWalker, conversions: Conversions
 ) : DachaClientServiceRegistry {
-  private val apiKeyService = DbDachaSqlApi(cacheSourceFeatureGroup, featureModelWalker, conversions)
+  private val apiKeyService = DbDachaSqlApi(cacheSourceFeatureGroup, featureModelWalker, commonCacheGenerator, conversions)
 
   override fun getEnvironmentService(cache: String): DachaEnvironmentService {
     return FakeDachaEnvironmentService() // is never called anyway
